Create Human Gait Analysis project
#**1. Human Gait Analysis project**

#Introduction
The goal of this project is to analyze human gait using machine learning techniques. We will be using a dataset that contains various features related to gait, such as step length, cadence, and stride time. The dataset was created by collecting data from a group of participants walking at different speeds and on different surfaces.
#Dataset Creation
To create a dataset for this project, we will use a hypothetical scenario where we are collecting gait data from 10 participants (5 male, 5 female) walking on a flat surface at a self-selected speed. We will use a smartphone app to collect the data, which will provide us with acceleration data for the x, y, and z axes. We will then use this data to calculate the following features:
*	Step length: the distance traveled by the foot from one heel strike to the next
*	Cadence: the number of steps taken per minute
*	Stride time: the time it takes to complete one full gait cycle (heel strike to heel strike)


We will then manually label each sample as either "normal" or "abnormal" based on visual inspection of the acceleration data.
Data Preparation
After collecting and labeling the data, we will split it into training and testing sets. We will use 70% of the data for training and 30% for testing. We will then normalize the features using Z-score normalization to ensure that they have the same scale.
#Machine Learning Model
We will use a simple binary classification model to classify each gait sample as either normal or abnormal. The model will consist of a single hidden layer with 10 nodes and a sigmoid activation function. We will use binary cross-entropy loss and the Adam optimizer.
#Code:
Below is the code for this project, written in Python using the TensorFlow library.