Create Automated Detection of Bone Fractures in X-ray Images using Deep Learning
Bone fractures are a common injury that requires immediate medical attention. X-ray images are the most commonly used diagnostic tool for detecting bone fractures. However, the interpretation of X-ray images is a time-consuming and subjective task that requires the expertise of a radiologist. In this project, the goal is to develop an automated system for detecting bone fractures in X-ray images using deep learning techniques. The system will use image processing techniques such as image segmentation, feature extraction, and convolutional neural networks (CNNs) to classify X-ray images as either normal or abnormal (i.e., containing a fracture). The Python libraries such as OpenCV, TensorFlow, and Keras can be used to preprocess and analyze the X-ray images, while libraries such as Matplotlib and Seaborn can be used to visualize the results. This project has the potential to significantly improve the efficiency and accuracy of bone fracture diagnosis, particularly in regions where access to radiologists is limited.